+++ test/rpc/test_session_manager.rb
@@ -570,6 +570,39 @@ class TestRPCSessionManager < KwardTestCase
     end
   end
 
+  def test_rpc_reload_starts_and_clears_plugin_footer
+    Dir.mktmpdir do |config_dir|
+      Dir.mktmpdir do |home|
+        plugins_dir = File.join(home, ".kward", "plugins")
+        plugin_path = File.join(plugins_dir, "footer.rb")
+        FileUtils.mkdir_p(plugins_dir)
+        server = RecordingServer.new
+        manager = Kward::RPC::SessionManager.new(server: server, client: FakeClient.new([]), config_dir: config_dir)
+        session = nil
+
+        with_env("HOME" => home, "KWARD_CONFIG_PATH" => nil) do
+          session = manager.create_session(workspace_root: Dir.pwd)
+          File.write(plugin_path, <<~'RUBY')
+            Kward.plugin do |plugin|
+              plugin.footer do |_ctx|
+                "Reloaded footer"
+              end
+            end
+          RUBY
+
+          manager.reload_plugins
+          File.delete(plugin_path)
+          manager.reload_plugins
+        end
+
+        footer_notifications = server.notifications.select { |notification| notification[:method] == "ui/footer" }
+        assert_equal({ sessionId: session[:id], text: "Reloaded footer" }, footer_notifications.first[:params])
+        assert_equal({ sessionId: session[:id], text: "" }, footer_notifications.last[:params])
+        manager.close_session(session_id: session[:id])
+      end
+    end
+  end
+
   def test_rpc_plugin_footer_refreshes_on_interval
     original_interval = Kward::RPC::SessionManager::FOOTER_REFRESH_INTERVAL
     Kward::RPC::SessionManager.send(:remove_const, :FOOTER_REFRESH_INTERVAL)